# docs/swagger.yaml
basePath: /api/v1
definitions:
  model.CreateConfigRequest:
    properties:
      key:
        type: string
      remark:
        type: string
      type:
        enum:
        - string
        - number
        - boolean
        - json
        type: string
      value:
        type: string
    required:
    - key
    - type
    - value
    type: object
...
  model.LoginRequest:
    properties:
      password:
        type: string
      phone:
        type: string
    required:
    - password
    - phone
    type: object
  model.RegisterRequest:
    properties:
      name:
        type: string
      password:
        minLength: 6
        type: string
      phone:
        type: string
    required:
    - name
    - password
    - phone
    type: object
  model.UpdateConfigRequest:
    properties:
      remark:
        type: string
      type:
        enum:
        - string
        - number
        - boolean
        - json
        type: string
      value:
        type: string
    required:
    - type
    - value
    type: object
...
  model.UpdateProfileRequest:
    properties:
      avatar:
        type: string
      name:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: 滴滴打人服务 API 文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: 滴滴打人 API
  version: "1.0"
paths:
...
  /login:
    post:
      consumes:
      - application/json
      description: 用户登录并获取token
      parameters:
      - description: 用户登录信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 用户登录
      tags:
      - 用户
  /register:
    post:
      consumes:
      - application/json
      description: 注册新用户
      parameters:
      - description: 用户注册信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 用户注册
      tags:
      - 用户
...
  /system/config:
    post:
      consumes:
      - application/json
      description: 创建新的系统配置项
      parameters:
      - description: 配置信息
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/model.CreateConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 创建系统配置
      tags:
      - 系统配置
  /system/config/{key}:
    delete:
      consumes:
      - application/json
      description: 删除指定的系统配置
      parameters:
      - description: 配置key
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 删除系统配置
      tags:
      - 系统配置
    get:
      consumes:
      - application/json
      description: 根据key获取系统配置
      parameters:
      - description: 配置key
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取系统配置
      tags:
      - 系统配置
    put:
      consumes:
      - application/json
      description: 更新系统配置信息
      parameters:
      - description: 配置key
        in: path
        name: key
        required: true
        type: string
      - description: 配置信息
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/model.UpdateConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 更新系统配置
      tags:
      - 系统配置
  /system/configs:
    get:
      consumes:
      - application/json
      description: 获取所有系统配置列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取所有系统配置
      tags:
      - 系统配置
  /user/profile:
    get:
      consumes:
      - application/json
      description: 获取当前登录用户的详细信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取用户信息
      tags:
      - 用户
    put:
      consumes:
      - application/json
      description: 更新当前登录用户的信息
      parameters:
      - description: 用户信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 更新用户信息
      tags:
      - 用户
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"